api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Movie Night Assistant API - Chat backend powered by Azure OpenAI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.12.0",
    "langchain-openai>=1.1.7",
    "langchain-core>=1.2.9",
    "langgraph>=0.4.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
addopts = ["--import-mode=prepend"]
